Start implement API server
This is a start on implementing an API server accessible with ZMQ over
TCP with the intention to be able to control the GUI remotely, e.g.
from a Tango server.
I first tried fastAPI with uvicorn to implement a RESET API but
somehow this confliceted with Qt (event-loop) and I couldn't make
it work. The best try was to start uvicorn in a seperated thread as
suggested here
https://stackoverflow.com/questions/61577643/how-to-use-fastapi-and-uvicorn-run-without-blocking-the-thread/64521239#64521239
but no luck. Another promissing approach is the QtHttpServer, which is
unfortunately still experimental and seems only be available for C++
at this point in time. Searching for "Qt rest api" and similiar didn't
result in anything helpful.
This crux is eventually going to have to go to pymepix, where it
should become a proper REST API or similar.